Don't trap JVM exits because the SecurityManager is deprecated
The JDK deprecated the SecurityManager: https://openjdk.org/jeps/411

In JDK 18+ accessing the SecurityManager throws an
UnsupportedOperationException.

There is no alternative way to trap JVM exits and prevent them. There is
a ticket for the JDK to add a way to do this
https://bugs.openjdk.org/browse/JDK-8199704

sbt ran into this same issue and resovled it just by not trapping exits
anymore
https://github.com/sbt/sbt/pull/6665
https://github.com/sbt/sbt/issues/6558