Use Zinc's new ConsistentAnalysisStore
This is new analysis store format added to Zinc by Databricks that is
deterministic. Given two identical Zinc states (after applying the read
write wappers) the outputs should be identical.

As an added bonus it is faster and smaller than the previous format. See
this PR for more info: https://github.com/sbt/zinc/pull/1326

This means we can stop most of the work we're doing to make the Zinc
analysis output more deterministic and just rely on this new analysis
format.