Temporarily add text analysis as additional output to find reproducibility problems
Consistent text format deserialization has a problem, so that's why I'm
adding the additional output for now. Will remove it once I've figured
out what is going on.

This is also very useful for debugging non-determinism issues in the
analysis store files, so I'm going to leave it enabled until we can
gather more data and be confident those are not problematic.