Workaround Zinc's libraryClassName non-determinism
libraryClassName in Zinc appears to be non-deterministic because you get
a jar mapped to a single class. I'm guessing that something isn't sorted
because the class you get isn't always the same.

We should fix the source of non-determinism in Zinc and then stop
filtering out the libraryClassName. Until then, we filter it from the
analysis store output.