Use an AnnexScalaInstance cache to prevent non-determinism in analysis store files
There's a class loader cache in Zinc and the Scala compiler also has a
some caching for its class path. Using those caches without a cache for
our AnnecScalaInstance (which is where we create our class loaders for
compilation), results in non-determinism in the analysis store files.

See the comment in AnnexScalaInstance for a more in depth explanation of
this. I don't fully understand it, but the empirical data I've gathered
so far backs up what is written in that comment.