Add support for multiplex sandboxing and improve argument parsing abstraction
These could/should probably be two separate commits, but the improvement
to argument parsing was inspired by the requirements of multiplex
sandboxing. The work was so intertwined that I'm leaving it as one PR.

Good news is that if we need to move to another parsing framework in the
future, it should be much easier because parsing is no longer spread
throughout the work functions.