Compare short paths in deps checker to avoid false positives from different configurations
Currently we compare the absolute path in the deps checker. This causes
us problems when there are configuration specific parts of one input and
not the other. This happens when there we've written a configuration
specific part of a path to disk and are comparing it to a path mapped
argument.

I imagine this is not the only place we'll run into issues with this.