Don't use @bazel_tools//tools/jdk:toolchain_utils.bzl
This file just contains aliases to
`@rules_java//toolchains:toolchain_utils.bzl`, so it's not that useful
anyway.

More importantly, Stardoc complains if we do because it needs a
`bzl_library` target for every `*.bzl` file we directly or indirectly
depend upon, and bazel_tools uses a different version of rules_java than
we do. This means we can't depend on `rules_java`'s `bzl_library`
target because Stardoc recognizes our version as a different repository.