Compute the SemanticDB target root correctly
In Bazel, `File#short_path` is prefixed with `..` if the file belongs to
an external repository. However, this isn't meant to be interpreted as a
relative path, as external repositories' outputs are in the `external`
output directory.

This meant that when `semanticdb_bundle` was enabled and external
repositories contained Scala rules, they'd fail to build, since the
`-P:semanticdb:sourceroot` or `-semanticdb-target` compiler options
would be computed incorrectly.