Don't depend on @annex//... in rules_scala_annex_test
This commit redeclares every dependency in `@annex//...` depended on in
`/tests` in the `annex_test` repository.

Fortunately in most cases, but unfortunately in this case, Bzlmod
prohibits repositories from depending on each other's external
repositories. That means that the repository in `/tests` can't depend on
anything in `@annex//...` unless it defines an `annex` repository
itself.

I considered using `include` so that we could declare `annex` with the
same set of dependencies in both repositories, but unfortunately,
`/tests` wouldn't be able to `include` such a file because, according to
the documentation for `include`, "only files in the main repo may be
included".