Use the new protobuf rules
Bazel 8 deprecates the rulesets bundled with Bazel (which includes the
Protobuf rules) in favor of their Starlark-ified equivalents. For that
reason, we switch to the `protobuf` ruleset for `proto_library` and
`java_proto_library`.

In doing so, we can't use `com.github.os72:protoc-jar` because the
version of protoc it bundles is extremely out of date and incompatible
with the version of protobuf provided by the Starlark-ified protobuf
ruleset. In its place, we use the protoc provided by that ruleset.