Don't generate SemanticDB for generated or external sources
Previously, we only checked if a source ended with the "scala" extension
before adding it to `SemanticDbInfo`. However, we don't want to generate
SemanticDB files for generated files or external files, because they'll
have the output directory (beginning with `bazel-out`) in their paths,
which isn't guaranteed to be consistent.

Also, tell the SemanticDB compiler plugin to generate SemanticDB outside
of the output JAR, even if we don't want SemanticDB generated for any of
the sources. That way, the output JAR won't include SemanticDB files we
don't want.